Unlimited Param Commands (only works for binary commands)

See merge request !32

+++ src/model/command/zeroParameter/turtle/ID.java
@@ -4,6 +4,13 @@ import java.util.List;
 import model.command.AbstractCommand;
 
 
+
+/**
+ * 
+ * @author Michael Schroeder
+ *
+ */
+
 public class ID extends AbstractCommandZeroParameterTurtle {
 
     public ID (List<AbstractCommand> parameters) {
@@ -13,7 +20,7 @@ public class ID extends AbstractCommandZeroParameterTurtle {
     @Override
     protected double turtleExecute () {
         System.out.println(myTurtle.getID() +" ID");
-        return myTurtle.getID();
+        return (double)myTurtle.getID();
     }
 
 }
